Fixed some errors
allowed the Thread.sleep to take long values to prevent overflows
Let Note allow 0 durations and fixed  tempo error in zoot-zl.txt

+++ src/cs3500/music/view/MIDIView.java
@@ -69,7 +69,7 @@ public class MIDIView implements IView {
       }
     }
     try {
-      Thread.sleep(controller.lastBeat() * controller.getTempo());
+      Thread.sleep((long)controller.lastBeat() * (long)controller.getTempo());
     } catch(InterruptedException e) {
       e.printStackTrace();
     }
